# perf: improve paragragh performance