--- a/lorem_test.go
+++ b/lorem_test.go
@@ -52,6 +52,7 @@ func TestFakeWord(t *testing.T) {
 
 func TestFakeSentence(t *testing.T) {
 	res := Sentence()
+	fmt.Println(res)
 	if res == "" || !strings.HasSuffix(res, ".") {
 		t.Error("Expected sentence")
 	}
@@ -59,6 +60,7 @@ func TestFakeSentence(t *testing.T) {
 
 func TestFakeParagraph(t *testing.T) {
 	res := Paragraph()
+	fmt.Println(res)
 	if res == "" || !strings.HasSuffix(res, ".") {
 		t.Error("Expected paragraph")
 	}
@@ -107,5 +109,11 @@ func TestUniqueParagraph(t *testing.T) {
 	ResetUnique()
 	if res == "" || !strings.HasSuffix(res, ".") {
 		t.Error("Expected paragraph")
+	}
+}
+
+func BenchmarkParagraph(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		_ = Paragraph()
 	}
 }