### summary

improve paragragh performance.

#### before optimize

```
goos: darwin
goarch: amd64
pkg: github.com/go-faker/faker/v4
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkParagraph
BenchmarkParagraph-12    	   21318	     50666 ns/op	   20836 B/op	     164 allocs/op
PASS
ok  	github.com/go-faker/faker/v4	1.833s
```

#### after optimize

```
goos: darwin
goarch: amd64
pkg: github.com/go-faker/faker/v4
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkParagraph
BenchmarkParagraph-12    	   25045	     43137 ns/op	   17784 B/op	      94 allocs/op
PASS
ok  	github.com/go-faker/faker/v4	1.867s
```